app-dep/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13.1
psycopg2>=2.9.3
django-cors-headers>=3.10.0
gunicorn>=20.1.0
python-dotenv>=0.21.0
dotenv
requests
python-dateutil
djangorestframework-simplejwt
cryptography